src/uv_project/pyproject.toml
[project]
name = "gsc"
version = "0.1.0"
description = "GSC dashboard and LoRa/plotting scripts for Torito"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pynput>=1.8.1",
    "smbus2>=0.6.0",
    "matplotlib>=3.5",
    "numpy>=1.20",
    "pandas>=1.3",
    "pyserial>=3.5",
]

[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'aarch64'"
]
